Изменение типов айди резервуаров в дто Pipe

+++ Simulation/src/main/java/ru/noxly/simulation/models/models/requests/PipeCreateDtoReq.java
@@ -11,7 +11,7 @@ public class PipeCreateDtoReq {
 
     private final Double diameter;
 
-    private final Integer sourceId;
+    private final Long sourceId;
 
-    private final Integer targetId;
+    private final Long targetId;
 }
